Adaugat autentificare cu Identity, servie, unit of work part 2

+++ Repositories/ArticolRepository.cs
@@ -1,62 +1,109 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Proiect.ContextModels;
 using Proiect.Entities;
+using Proiect.Services;
 using Proiect.Models;
-using Microsoft.VisualBasic;
 
 namespace Proiect.Repositories
 {
     public class ArticolRepository: IArticolRepository
     {
         private readonly ProiectContext _context;
-        public ArticolRepository(ProiectContext context)
+        private readonly IUtilizatorService _utilizatorService;
+        public ArticolRepository(ProiectContext context, IUtilizatorService utilizatorService)
         {
             _context = context;
+            _utilizatorService = utilizatorService;
         }
         public async Task<ICollection<Articol>> GetArticoleAsync()
         {
             return (await _context.Articol.ToListAsync());
         }
-        public async Task<Articol?> GetArticolAsync(int id)
+        public async Task<Articol?> GetArticolAsync(string titlu)
         {
-            var articol = await _context.Articol.FindAsync(id);
-            return articol;
+            var articol = await _context.Articol.Where(a => a.Titlu == titlu).ToListAsync(); 
+
+            if(articol == null)
+                return null;
+
+            return articol[0];
         }
-        public async Task<ICollection<Articol>?> GetArticolAutorAsync(string id)
+        public async Task<ICollection<Articol>?> GetArticolAutorAsync(string userName)
         {
-            var articol = await _context.Articol.Where(x => x.UtilizatorId == id).
+            var articol = await _context.Articol.Where(x => x.Utilizator.UserName == userName).
                 OrderBy(x => x.Titlu).ToListAsync();
+
             return articol;
         }
         public async Task<List<ArticolUtilizatorDto>?> GetArticoleGrupateAsync()
         {
             var articole = await _context.Articol
-                        .GroupBy(a => a.UtilizatorId)
-                        .OrderBy(a => a.Key)
+                        .GroupBy(a => new { a.UtilizatorId, a.Utilizator.UserName })
+                        .OrderBy(a => a.Key.UtilizatorId)
                         .Select(grup => new ArticolUtilizatorDto  // Fiecare grup o sa fie reprezentat de un UtilizatorId si o                       
                         {                                    // lista de articole ordonate dupa titlu
-                            UtilizatorId = grup.Key,
+                            UserName = grup.Key.UserName,
                             Articole = grup.OrderBy(a => a.Titlu).ToList()
                         })
                         .ToListAsync();
 
             return articole;
         }
-        public async Task<Articol> PutArticolAsync(Articol articol)
+        public async Task<Articol?> PutArticolAsync(string? userName, int id, Articol articol)
         {
-            _context.Articol.Update(articol);
+            // Verifica daca exista articolul si daca este scris de utilizatorul care face request ul
+
+            var ar = await _context.Articol.FindAsync(id);
+
+            if (ar == null)
+                return null;
+
+            if (!await _utilizatorService.VerificaArticol(userName, ar))
+                return null;
+
+            ar.Titlu = articol.Titlu;
+            ar.Continut = articol.Continut;
+
+            _context.Articol.Update(ar);
             await _context.SaveChangesAsync();
-            return articol;
+
+            return ar;
         }
-        public async Task PostArticolAsync(Articol articol)
+        public async Task<Articol?> PostArticolAsync(string userName, Articol articol)
         {
-            _context.Articol.Add(articol);
+            var ar = await _utilizatorService.AdaugaArticol(userName, articol);
+            
+            if(ar == null)
+                return null;
+
+            _context.Articol.Add(ar);
             await _context.SaveChangesAsync();
+
+            return ar;
         }
-        public async Task DeleteArticolAsync(Articol articol)
+        public async Task<bool> DeleteArticolAsync(int id)
         {
+            var articol = await _context.Articol.FindAsync(id);
+
+            if (articol == null)
+                return false;
+
             _context.Articol.Remove(articol);
             await _context.SaveChangesAsync();
+
+            return true;
+        }
+        public async Task<bool> DeleteArticolUtilizatorAsync(string userName, Articol articol)
+        {
+            if(await _utilizatorService.VerificaArticol(userName, articol))
+            {
+                _context.Articol.Remove(articol);
+                await _context.SaveChangesAsync();
+
+                return true;
+            }
+
+            return false;
         }
     }
 }
